refactor: move ToolSettings and ToolTimeoutResolver to config_resolvers package
