+++ src/quickapp/config_resolvers/tool_timeout_resolver.py
@@ -1,7 +1,7 @@
 from injector import inject
 
-from quickapp.common.tool_timeout_settings import ToolSettings
 from quickapp.config.application import ApplicationConfig
+from quickapp.config_resolvers.tool_timeout_settings import ToolSettings
 
 
 @inject
