+++ src/quickapp/dial_deployment_tooling/dial_completion_service.py
@@ -24,7 +24,7 @@ from quickapp.common.chat_completion_stream.stream_result import ChatStreamAccum
 from quickapp.common.deployment_usage import DeploymentUsage
 from quickapp.common.dial_settings import DialSettings
 from quickapp.common.file_reference_pattern import strip_file_prefix
-from quickapp.common.tool_timeout_resolver import ToolTimeoutResolver
+from quickapp.config_resolvers.tool_timeout_resolver import ToolTimeoutResolver
 from quickapp.common.tool_timeout_utils import translate_timeout
 from quickapp.dial_deployment_tooling.constants import (
     ATTACHMENT_PARAM,
