+++ src/quickapp/mcp_tooling/_mcp_connection_manager.py
@@ -13,7 +13,7 @@ from mcp.types import CallToolResult
 from quickapp.common import DIAL_BEARER, ForwardedHeaders
 from quickapp.common.dial_settings import DialSettings
 from quickapp.common.oauth_token_fetcher import OAuthTokenFetcher
-from quickapp.common.tool_timeout_resolver import ToolTimeoutResolver
+from quickapp.config_resolvers.tool_timeout_resolver import ToolTimeoutResolver
 from quickapp.common.tool_timeout_utils import MCP_TIMEOUT_CODE
 from quickapp.config.toolsets.authorization import (
     BasicAuthorization,
